drink-selector-main/requirements.txt
flask
gunicorn
reportlab

google-auth
google-auth-oauthlib
google-auth-httplib2




